Revert "HDDS-1395. Key write fails with BlockOutputStream has been closed exception (#749)"
This reverts commit d525633619f83a76bcb4d73b444bbfe8e3ab594b.
