+++ hadoop-ozone/client/src/main/java/org/apache/hadoop/ozone/client/io/KeyOutputStream.java
@@ -23,18 +23,21 @@ import org.apache.hadoop.fs.FSExceptionMessages;
 import org.apache.hadoop.fs.FileEncryptionInfo;
 import org.apache.hadoop.hdds.protocol.DatanodeDetails;
 import org.apache.hadoop.hdds.protocol.datanode.proto.ContainerProtos.ChecksumType;
-import org.apache.hadoop.hdds.scm.client.HddsClientUtils;
 import org.apache.hadoop.hdds.scm.container.ContainerID;
 import org.apache.hadoop.hdds.scm.container.common.helpers.ContainerNotOpenException;
+import org.apache.hadoop.hdds.scm.storage.BufferPool;
+import org.apache.hadoop.ozone.OzoneConfigKeys;
 import org.apache.hadoop.hdds.scm.container.common.helpers.ExcludeList;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineID;
 import org.apache.hadoop.io.retry.RetryPolicies;
 import org.apache.hadoop.io.retry.RetryPolicy;
+import org.apache.hadoop.ozone.client.OzoneClientUtils;
 import org.apache.hadoop.ozone.om.helpers.*;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos.ReplicationType;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos.ReplicationFactor;
 import org.apache.hadoop.ozone.om.protocol.OzoneManagerProtocol;
 import org.apache.hadoop.hdds.scm.XceiverClientManager;
+import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.ratis.protocol.AlreadyClosedException;
 import org.apache.ratis.protocol.GroupMismatchException;
 import org.apache.ratis.protocol.RaftRetryFailureException;
@@ -44,8 +47,10 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.io.InterruptedIOException;
 import java.io.OutputStream;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Collection;
+import java.util.ListIterator;
 import java.util.Map;
 import java.util.concurrent.TimeoutException;
 import java.util.function.Function;
@@ -72,41 +77,84 @@ public class KeyOutputStream extends OutputStream {
   public static final Logger LOG =
       LoggerFactory.getLogger(KeyOutputStream.class);
 
+  // array list's get(index) is O(1)
+  private final ArrayList<BlockOutputStreamEntry> streamEntries;
+  private int currentStreamIndex;
+  private final OzoneManagerProtocol omClient;
+  private final OmKeyArgs keyArgs;
+  private final long openID;
+  private final XceiverClientManager xceiverClientManager;
+  private final int chunkSize;
+  private final String requestID;
   private boolean closed;
+  private final long streamBufferFlushSize;
+  private final long streamBufferMaxSize;
+  private final long watchTimeout;
+  private final long blockSize;
+  private final int bytesPerChecksum;
+  private final ChecksumType checksumType;
+  private final BufferPool bufferPool;
+  private OmMultipartCommitUploadPartInfo commitUploadPartInfo;
   private FileEncryptionInfo feInfo;
+  private ExcludeList excludeList;
   private final Map<Class<? extends Throwable>, RetryPolicy> retryPolicyMap;
   private int retryCount;
   private long offset;
-  private final BlockOutputStreamEntryPool blockOutputStreamEntryPool;
-
   /**
    * A constructor for testing purpose only.
    */
   @VisibleForTesting
+  @SuppressWarnings("parameternumber")
   public KeyOutputStream() {
+    streamEntries = new ArrayList<>();
+    omClient = null;
+    keyArgs = null;
+    openID = -1;
+    xceiverClientManager = null;
+    chunkSize = 0;
+    requestID = null;
     closed = false;
-    this.retryPolicyMap = HddsClientUtils.getExceptionList()
+    streamBufferFlushSize = 0;
+    streamBufferMaxSize = 0;
+    bufferPool = new BufferPool(chunkSize, 1);
+    watchTimeout = 0;
+    blockSize = 0;
+    this.checksumType = ChecksumType.valueOf(
+        OzoneConfigKeys.OZONE_CLIENT_CHECKSUM_TYPE_DEFAULT);
+    this.bytesPerChecksum = OzoneConfigKeys
+        .OZONE_CLIENT_BYTES_PER_CHECKSUM_DEFAULT_BYTES; // Default is 1MB
+    this.retryPolicyMap = OzoneClientUtils.getExceptionList()
         .stream()
         .collect(Collectors.toMap(Function.identity(),
             e -> RetryPolicies.TRY_ONCE_THEN_FAIL));
     retryCount = 0;
     offset = 0;
-    blockOutputStreamEntryPool = new BlockOutputStreamEntryPool();
   }
 
   @VisibleForTesting
   public List<BlockOutputStreamEntry> getStreamEntries() {
-    return blockOutputStreamEntryPool.getStreamEntries();
+    return streamEntries;
   }
-
   @VisibleForTesting
   public XceiverClientManager getXceiverClientManager() {
-    return blockOutputStreamEntryPool.getXceiverClientManager();
+    return xceiverClientManager;
   }
 
-  @VisibleForTesting
-  public List<OmKeyLocationInfo> getLocationInfoList() {
-    return blockOutputStreamEntryPool.getLocationInfoList();
+  public List<OmKeyLocationInfo> getLocationInfoList() throws IOException {
+    List<OmKeyLocationInfo> locationInfoList = new ArrayList<>();
+    for (BlockOutputStreamEntry streamEntry : streamEntries) {
+      OmKeyLocationInfo info =
+          new OmKeyLocationInfo.Builder().setBlockID(streamEntry.getBlockID())
+              .setLength(streamEntry.getCurrentPosition()).setOffset(0)
+              .setToken(streamEntry.getToken())
+              .setPipeline(streamEntry.getPipeline())
+              .build();
+      LOG.debug("block written " + streamEntry.getBlockID() + ", length "
+          + streamEntry.getCurrentPosition() + " bcsID "
+          + streamEntry.getBlockID().getBlockCommitSequenceId());
+      locationInfoList.add(info);
+    }
+    return locationInfoList;
   }
 
   @VisibleForTesting
@@ -123,16 +171,41 @@ public class KeyOutputStream extends OutputStream {
       ChecksumType checksumType, int bytesPerChecksum,
       String uploadID, int partNumber, boolean isMultipart,
       int maxRetryCount, long retryInterval) {
+    this.streamEntries = new ArrayList<>();
+    this.currentStreamIndex = 0;
+    this.omClient = omClient;
     OmKeyInfo info = handler.getKeyInfo();
-    blockOutputStreamEntryPool =
-        new BlockOutputStreamEntryPool(omClient, chunkSize, requestId, factor,
-            type, bufferFlushSize, bufferMaxSize, size, watchTimeout,
-            checksumType, bytesPerChecksum, uploadID, partNumber, isMultipart,
-            info, xceiverClientManager, handler.getId());
     // Retrieve the file encryption key info, null if file is not in
     // encrypted bucket.
     this.feInfo = info.getFileEncryptionInfo();
-    this.retryPolicyMap = HddsClientUtils.getRetryPolicyByException(
+    this.keyArgs = new OmKeyArgs.Builder().setVolumeName(info.getVolumeName())
+        .setBucketName(info.getBucketName()).setKeyName(info.getKeyName())
+        .setType(type).setFactor(factor).setDataSize(info.getDataSize())
+        .setIsMultipartKey(isMultipart).setMultipartUploadID(
+            uploadID).setMultipartUploadPartNumber(partNumber)
+        .build();
+    this.openID = handler.getId();
+    this.xceiverClientManager = xceiverClientManager;
+    this.chunkSize = chunkSize;
+    this.requestID = requestId;
+    this.streamBufferFlushSize = bufferFlushSize;
+    this.streamBufferMaxSize = bufferMaxSize;
+    this.blockSize = size;
+    this.watchTimeout = watchTimeout;
+    this.bytesPerChecksum = bytesPerChecksum;
+    this.checksumType = checksumType;
+
+    Preconditions.checkState(chunkSize > 0);
+    Preconditions.checkState(streamBufferFlushSize > 0);
+    Preconditions.checkState(streamBufferMaxSize > 0);
+    Preconditions.checkState(blockSize > 0);
+    Preconditions.checkState(streamBufferFlushSize % chunkSize == 0);
+    Preconditions.checkState(streamBufferMaxSize % streamBufferFlushSize == 0);
+    Preconditions.checkState(blockSize % streamBufferMaxSize == 0);
+    this.bufferPool =
+        new BufferPool(chunkSize, (int)streamBufferMaxSize / chunkSize);
+    this.excludeList = new ExcludeList();
+    this.retryPolicyMap = OzoneClientUtils.getRetryPolicyByException(
         maxRetryCount, retryInterval);
     this.retryCount = 0;
   }
@@ -152,7 +225,37 @@ public class KeyOutputStream extends OutputStream {
    */
   public void addPreallocateBlocks(OmKeyLocationInfoGroup version,
       long openVersion) throws IOException {
-    blockOutputStreamEntryPool.addPreallocateBlocks(version, openVersion);
+    // server may return any number of blocks, (0 to any)
+    // only the blocks allocated in this open session (block createVersion
+    // equals to open session version)
+    for (OmKeyLocationInfo subKeyInfo : version.getLocationList()) {
+      if (subKeyInfo.getCreateVersion() == openVersion) {
+        addKeyLocationInfo(subKeyInfo);
+      }
+    }
+  }
+
+  private void addKeyLocationInfo(OmKeyLocationInfo subKeyInfo)
+      throws IOException {
+    Preconditions.checkNotNull(subKeyInfo.getPipeline());
+    UserGroupInformation.getCurrentUser().addToken(subKeyInfo.getToken());
+    BlockOutputStreamEntry.Builder builder =
+        new BlockOutputStreamEntry.Builder()
+            .setBlockID(subKeyInfo.getBlockID())
+            .setKey(keyArgs.getKeyName())
+            .setXceiverClientManager(xceiverClientManager)
+            .setPipeline(subKeyInfo.getPipeline())
+            .setRequestId(requestID)
+            .setChunkSize(chunkSize)
+            .setLength(subKeyInfo.getLength())
+            .setStreamBufferFlushSize(streamBufferFlushSize)
+            .setStreamBufferMaxSize(streamBufferMaxSize)
+            .setWatchTimeout(watchTimeout)
+            .setbufferPool(bufferPool)
+            .setChecksumType(checksumType)
+            .setBytesPerChecksum(bytesPerChecksum)
+            .setToken(subKeyInfo.getToken());
+    streamEntries.add(builder.build());
   }
 
   @Override
@@ -191,12 +294,34 @@ public class KeyOutputStream extends OutputStream {
     handleWrite(b, off, len, false);
   }
 
+  private long computeBufferData() {
+    return bufferPool.computeBufferData();
+  }
+
   private void handleWrite(byte[] b, int off, long len, boolean retry)
       throws IOException {
+    int succeededAllocates = 0;
     while (len > 0) {
       try {
-        BlockOutputStreamEntry current =
-            blockOutputStreamEntryPool.allocateBlockIfNeeded();
+        if (streamEntries.size() <= currentStreamIndex) {
+          Preconditions.checkNotNull(omClient);
+          // allocate a new block, if a exception happens, log an error and
+          // throw exception to the caller directly, and the write fails.
+          try {
+            allocateNewBlock(currentStreamIndex);
+            succeededAllocates += 1;
+          } catch (IOException ioe) {
+            LOG.error("Try to allocate more blocks for write failed, already "
+                + "allocated " + succeededAllocates
+                + " blocks for this write.");
+            throw ioe;
+          }
+        }
+        // in theory, this condition should never violate due the check above
+        // still do a sanity check.
+        Preconditions.checkArgument(currentStreamIndex < streamEntries.size());
+        BlockOutputStreamEntry current = streamEntries.get(currentStreamIndex);
+
         // length(len) will be in int range if the call is happening through
         // write API of blockOutputStream. Length can be in long range if it
         // comes via Exception path.
@@ -217,8 +342,7 @@ public class KeyOutputStream extends OutputStream {
           // to or less than the max length of the buffer allocated.
           // The len specified here is the combined sum of the data length of
           // the buffers
-          Preconditions.checkState(!retry || len <= blockOutputStreamEntryPool
-              .getStreamBufferMaxSize());
+          Preconditions.checkState(!retry || len <= streamBufferMaxSize);
           int dataWritten = (int) (current.getWrittenDataLength() - currentPos);
           writeLen = retry ? (int) len : dataWritten;
           // In retry path, the data written is already accounted in offset.
@@ -226,7 +350,7 @@ public class KeyOutputStream extends OutputStream {
             offset += writeLen;
           }
           LOG.debug("writeLen {}, total len {}", writeLen, len);
-          handleException(current, ioe);
+          handleException(current, currentStreamIndex, ioe);
         }
         if (current.getRemaining() <= 0) {
           // since the current block is already written close the stream.
@@ -242,18 +366,79 @@ public class KeyOutputStream extends OutputStream {
   }
 
   /**
+   * Discards the subsequent pre allocated blocks and removes the streamEntries
+   * from the streamEntries list for the container which is closed.
+   * @param containerID id of the closed container
+   * @param pipelineId id of the associated pipeline
+   * @param streamIndex index of the stream
+   */
+  private void discardPreallocatedBlocks(long containerID,
+      PipelineID pipelineId, int streamIndex) {
+    // streamIndex < streamEntries.size() signifies that, there are still
+    // pre allocated blocks available.
+
+    // This will be called only to discard the next subsequent unused blocks
+    // in the streamEntryList.
+    if (streamIndex < streamEntries.size()) {
+      ListIterator<BlockOutputStreamEntry> streamEntryIterator =
+          streamEntries.listIterator(streamIndex);
+      while (streamEntryIterator.hasNext()) {
+        BlockOutputStreamEntry streamEntry = streamEntryIterator.next();
+        Preconditions.checkArgument(streamEntry.getCurrentPosition() == 0);
+        if (((pipelineId != null && streamEntry.getPipeline().getId()
+            .equals(pipelineId)) || (containerID != -1
+            && streamEntry.getBlockID().getContainerID() == containerID))) {
+          streamEntryIterator.remove();
+        }
+      }
+    }
+  }
+
+  /**
+   * It might be possible that the blocks pre allocated might never get written
+   * while the stream gets closed normally. In such cases, it would be a good
+   * idea to trim down the locationInfoList by removing the unused blocks if any
+   * so as only the used block info gets updated on OzoneManager during close.
+   */
+  private void removeEmptyBlocks() {
+    if (currentStreamIndex < streamEntries.size()) {
+      ListIterator<BlockOutputStreamEntry> streamEntryIterator =
+          streamEntries.listIterator(currentStreamIndex);
+      while (streamEntryIterator.hasNext()) {
+        if (streamEntryIterator.next().getCurrentPosition() == 0) {
+          streamEntryIterator.remove();
+        }
+      }
+    }
+  }
+
+  private void cleanup() {
+    if (excludeList != null) {
+      excludeList.clear();
+      excludeList = null;
+    }
+    if (bufferPool != null) {
+      bufferPool.clearBufferPool();
+    }
+
+    if (streamEntries != null) {
+      streamEntries.clear();
+    }
+  }
+  /**
    * It performs following actions :
    * a. Updates the committed length at datanode for the current stream in
-   * datanode.
+   *    datanode.
    * b. Reads the data from the underlying buffer and writes it the next stream.
    *
    * @param streamEntry StreamEntry
-   * @param exception   actual exception that occurred
+   * @param streamIndex Index of the entry
+   * @param exception actual exception that occurred
    * @throws IOException Throws IOException if Write fails
    */
   private void handleException(BlockOutputStreamEntry streamEntry,
-      IOException exception) throws IOException {
-    Throwable t = HddsClientUtils.checkForException(exception);
+      int streamIndex, IOException exception) throws IOException {
+    Throwable t = checkForException(exception);
     boolean retryFailure = checkForRetryFailure(t);
     boolean closedContainerException = false;
     if (!retryFailure) {
@@ -263,19 +448,15 @@ public class KeyOutputStream extends OutputStream {
     long totalSuccessfulFlushedData = streamEntry.getTotalAckDataLength();
     //set the correct length for the current stream
     streamEntry.setCurrentPosition(totalSuccessfulFlushedData);
-    long bufferedDataLen = blockOutputStreamEntryPool.computeBufferData();
-    LOG.debug(
-        "Encountered exception {}. The last committed block length is {}, "
+    long bufferedDataLen = computeBufferData();
+    LOG.warn("Encountered exception {}. The last committed block length is {}, "
             + "uncommitted data length is {} retry count {}", exception,
         totalSuccessfulFlushedData, bufferedDataLen, retryCount);
-    Preconditions.checkArgument(
-        bufferedDataLen <= blockOutputStreamEntryPool.getStreamBufferMaxSize());
-    Preconditions.checkArgument(
-        offset - blockOutputStreamEntryPool.getKeyLength() == bufferedDataLen);
+    Preconditions.checkArgument(bufferedDataLen <= streamBufferMaxSize);
+    Preconditions.checkArgument(offset - getKeyLength() == bufferedDataLen);
     long containerId = streamEntry.getBlockID().getContainerID();
     Collection<DatanodeDetails> failedServers = streamEntry.getFailedServers();
     Preconditions.checkNotNull(failedServers);
-    ExcludeList excludeList = blockOutputStreamEntryPool.getExcludeList();
     if (!failedServers.isEmpty()) {
       excludeList.addDatanodes(failedServers);
     }
@@ -289,42 +470,45 @@ public class KeyOutputStream extends OutputStream {
     // just clean up the current stream.
     streamEntry.cleanup(retryFailure);
 
-    // discard all subsequent blocks the containers and pipelines which
+    // discard all sunsequent blocks the containers and pipelines which
     // are in the exclude list so that, the very next retry should never
     // write data on the  closed container/pipeline
     if (closedContainerException) {
       // discard subsequent pre allocated blocks from the streamEntries list
       // from the closed container
-      blockOutputStreamEntryPool
-          .discardPreallocatedBlocks(streamEntry.getBlockID().getContainerID(),
-              null);
+      discardPreallocatedBlocks(streamEntry.getBlockID().getContainerID(), null,
+          streamIndex + 1);
     } else {
       // In case there is timeoutException or Watch for commit happening over
       // majority or the client connection failure to the leader in the
-      // pipeline, just discard all the pre allocated blocks on this pipeline.
+      // pipeline, just discard all the preallocated blocks on this pipeline.
       // Next block allocation will happen with excluding this specific pipeline
       // This will ensure if 2 way commit happens , it cannot span over multiple
       // blocks
-      blockOutputStreamEntryPool
-          .discardPreallocatedBlocks(-1, pipelineId);
+      discardPreallocatedBlocks(-1, pipelineId, streamIndex + 1);
     }
     if (bufferedDataLen > 0) {
       // If the data is still cached in the underlying stream, we need to
       // allocate new block and write this data in the datanode.
+      currentStreamIndex += 1;
       handleRetry(exception, bufferedDataLen);
       // reset the retryCount after handling the exception
       retryCount = 0;
     }
+    if (totalSuccessfulFlushedData == 0) {
+      streamEntries.remove(streamIndex);
+      currentStreamIndex -= 1;
+    }
   }
 
   private void markStreamClosed() {
-    blockOutputStreamEntryPool.cleanup();
+    cleanup();
     closed = true;
   }
 
   private void handleRetry(IOException exception, long len) throws IOException {
-    RetryPolicy retryPolicy = retryPolicyMap
-        .get(HddsClientUtils.checkForException(exception).getClass());
+    RetryPolicy retryPolicy =
+        retryPolicyMap.get(checkForException(exception).getClass());
     if (retryPolicy == null) {
       retryPolicy = retryPolicyMap.get(Exception.class);
     }
@@ -360,11 +544,10 @@ public class KeyOutputStream extends OutputStream {
       }
     }
     retryCount++;
-    LOG.trace("Retrying Write request. Already tried " + retryCount
-        + " time(s); retry policy is " + retryPolicy);
+    LOG.trace("Retrying Write request. Already tried "
+        + retryCount + " time(s); retry policy is " + retryPolicy);
     handleWrite(null, 0, len, true);
   }
-
   /**
    * Checks if the provided exception signifies retry failure in ratis client.
    * In case of retry failure, ratis client throws RaftRetryFailureException
@@ -377,6 +560,40 @@ public class KeyOutputStream extends OutputStream {
 
   private boolean checkIfContainerIsClosed(Throwable t) {
     return t instanceof ContainerNotOpenException;
+  }
+
+  public Throwable checkForException(IOException ioe) throws IOException {
+    Throwable t = ioe.getCause();
+    while (t != null) {
+      for (Class<? extends Exception> cls : OzoneClientUtils
+          .getExceptionList()) {
+        if (cls.isInstance(t)) {
+          return t;
+        }
+      }
+      t = t.getCause();
+    }
+    throw ioe;
+  }
+
+  private long getKeyLength() {
+    return streamEntries.stream().mapToLong(e -> e.getCurrentPosition())
+        .sum();
+  }
+
+  /**
+   * Contact OM to get a new block. Set the new block with the index (e.g.
+   * first block has index = 0, second has index = 1 etc.)
+   *
+   * The returned block is made to new BlockOutputStreamEntry to write.
+   *
+   * @param index the index of the block.
+   * @throws IOException
+   */
+  private void allocateNewBlock(int index) throws IOException {
+    OmKeyLocationInfo subKeyInfo =
+        omClient.allocateBlock(keyArgs, openID, excludeList);
+    addKeyLocationInfo(subKeyInfo);
   }
 
   @Override
@@ -395,19 +612,20 @@ public class KeyOutputStream extends OutputStream {
    * written to new stream , it will be at max half full. In such cases, we
    * should just write the data and not close the stream as the block won't be
    * completely full.
-   *
    * @param op Flag which decides whether to call close or flush on the
-   *           outputStream.
+   *              outputStream.
    * @throws IOException In case, flush or close fails with exception.
    */
   private void handleFlushOrClose(StreamAction op) throws IOException {
-    if (blockOutputStreamEntryPool.isEmpty()) {
+    if (streamEntries.size() == 0) {
       return;
     }
     while (true) {
       try {
-        BlockOutputStreamEntry entry =
-            blockOutputStreamEntryPool.getCurrentStreamEntry();
+        int size = streamEntries.size();
+        int streamIndex =
+            currentStreamIndex >= size ? size - 1 : currentStreamIndex;
+        BlockOutputStreamEntry entry = streamEntries.get(streamIndex);
         if (entry != null) {
           try {
             Collection<DatanodeDetails> failedServers =
@@ -415,8 +633,7 @@ public class KeyOutputStream extends OutputStream {
             // failed servers can be null in case there is no data written in
             // the stream
             if (failedServers != null && !failedServers.isEmpty()) {
-              blockOutputStreamEntryPool.getExcludeList()
-                  .addDatanodes(failedServers);
+              excludeList.addDatanodes(failedServers);
             }
             switch (op) {
             case CLOSE:
@@ -425,6 +642,7 @@ public class KeyOutputStream extends OutputStream {
             case FULL:
               if (entry.getRemaining() == 0) {
                 entry.close();
+                currentStreamIndex++;
               }
               break;
             case FLUSH:
@@ -434,7 +652,7 @@ public class KeyOutputStream extends OutputStream {
               throw new IOException("Invalid Operation");
             }
           } catch (IOException ioe) {
-            handleException(entry, ioe);
+            handleException(entry, streamIndex, ioe);
             continue;
           }
         }
@@ -459,16 +677,34 @@ public class KeyOutputStream extends OutputStream {
     closed = true;
     try {
       handleFlushOrClose(StreamAction.CLOSE);
-      blockOutputStreamEntryPool.commitKey(offset);
+      if (keyArgs != null) {
+        // in test, this could be null
+        removeEmptyBlocks();
+        long length = getKeyLength();
+        Preconditions.checkArgument(offset == length);
+        keyArgs.setDataSize(length);
+        keyArgs.setLocationInfoList(getLocationInfoList());
+        // When the key is multipart upload part file upload, we should not
+        // commit the key, as this is not an actual key, this is a just a
+        // partial key of a large file.
+        if (keyArgs.getIsMultipartKey()) {
+          commitUploadPartInfo = omClient.commitMultipartUploadPart(keyArgs,
+              openID);
+        } else {
+          omClient.commitKey(keyArgs, openID);
+        }
+      } else {
+        LOG.warn("Closing KeyOutputStream, but key args is null");
+      }
     } catch (IOException ioe) {
       throw ioe;
     } finally {
-      blockOutputStreamEntryPool.cleanup();
+      cleanup();
     }
   }
 
   public OmMultipartCommitUploadPartInfo getCommitUploadPartInfo() {
-    return blockOutputStreamEntryPool.getCommitUploadPartInfo();
+    return commitUploadPartInfo;
   }
 
   public FileEncryptionInfo getFileEncryptionInfo() {
@@ -477,7 +713,7 @@ public class KeyOutputStream extends OutputStream {
 
   @VisibleForTesting
   public ExcludeList getExcludeList() {
-    return blockOutputStreamEntryPool.getExcludeList();
+    return excludeList;
   }
 
   /**
@@ -503,6 +739,7 @@ public class KeyOutputStream extends OutputStream {
     private int maxRetryCount;
     private long retryInterval;
 
+
     public Builder setMultipartUploadID(String uploadID) {
       this.multipartUploadID = uploadID;
       return this;
@@ -523,7 +760,8 @@ public class KeyOutputStream extends OutputStream {
       return this;
     }
 
-    public Builder setOmClient(OzoneManagerProtocol client) {
+    public Builder setOmClient(
+        OzoneManagerProtocol client) {
       this.omClient = client;
       return this;
     }
@@ -568,12 +806,12 @@ public class KeyOutputStream extends OutputStream {
       return this;
     }
 
-    public Builder setChecksumType(ChecksumType cType) {
+    public Builder setChecksumType(ChecksumType cType){
       this.checksumType = cType;
       return this;
     }
 
-    public Builder setBytesPerChecksum(int bytes) {
+    public Builder setBytesPerChecksum(int bytes){
       this.bytesPerChecksum = bytes;
       return this;
     }
@@ -593,9 +831,9 @@ public class KeyOutputStream extends OutputStream {
       return this;
     }
 
-    public KeyOutputStream build() {
-      return new KeyOutputStream(openHandler, xceiverManager, omClient,
-          chunkSize, requestID, factor, type, streamBufferFlushSize,
+    public KeyOutputStream build() throws IOException {
+      return new KeyOutputStream(openHandler, xceiverManager,
+          omClient, chunkSize, requestID, factor, type, streamBufferFlushSize,
           streamBufferMaxSize, blockSize, watchTimeout, checksumType,
           bytesPerChecksum, multipartUploadID, multipartNumber, isMultipartKey,
           maxRetryCount, retryInterval);
@@ -610,8 +848,8 @@ public class KeyOutputStream extends OutputStream {
   private void checkNotClosed() throws IOException {
     if (closed) {
       throw new IOException(
-          ": " + FSExceptionMessages.STREAM_IS_CLOSED + " Key: "
-              + blockOutputStreamEntryPool.getKeyName());
+          ": " + FSExceptionMessages.STREAM_IS_CLOSED + " Key: " + keyArgs
+              .getKeyName());
     }
   }
 }
